task-3-demo/requirements.txt
langchain-core
langchain-community
pandas
openai
python-dotenv
mcp
langgraph
langchain-mcp-adapters
langchain-openai